LayoutTests/media/picture-in-picture/picture-in-picture-api-exit-pip-2-expected.txt

This tests that exit Picture-in-Picture rejects when there is no Picture-in-Picture video.

RUN(internals.settings.setAllowsPictureInPictureMediaPlayback(true))
RUN(internals.setMockVideoPresentationModeEnabled(true))
EXPECTED (error.name == 'InvalidStateError') OK
END OF TEST
